Move ibis.param() to api.py
